reference-app/backend/requirements.txt
Flask==1.1.2
requests==2.23.0
pylint==2.4.0
gunicorn==20.0.4
pymongo==3.11.0
Flask-PyMongo==2.3.0
jaeger-client==4.3.0
prometheus-client==0.12.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-jaeger
prometheus-flask-exporter